Move upload processing code into service
The private fields required to keep track of temporary files, progress, etc confused the django ORM. So fields related to processing were moved out of the model object and into a separate service that operates on the model.

Updated the upload-unparsed script to work with the service.
The upload-lassy script looks like it should still work as it doesn't use the Upload model. But this will have to be verified.