Add support for user accounts
Doesn't do anything yet in the regular flow, but required in for users to be able to upload and manage treebanks in the future.
Currently it's not possible to configure the settings using environment variables when running in docker. Integration with social providers/institutional login is also not currently implemented (though allauth does support this).