Add frontend container for live reloading. Useful for development.
Running docker compose up -d frontend-dev will create a container that exposes a live-reloading server on port 4200. Installing of dependencies etc is take care of automatically.
The frontend will forward all requests to the backend to the "gretel-backend" container.
To start that container, run either "compose up backend" for a production container, or "compose up backend-dev" for a similarly live-reloading backend container.